+++ ErsatzTV/ViewModels/FFmpegProfileEditViewModel.cs
@@ -35,6 +35,7 @@ public class FFmpegProfileEditViewModel
         VideoFormat = viewModel.VideoFormat;
         VideoProfile = viewModel.VideoProfile;
         VideoPreset = viewModel.VideoPreset;
+        AllowBFrames = viewModel.AllowBFrames;
         BitDepth = viewModel.BitDepth;
     }
 
@@ -60,6 +61,7 @@ public class FFmpegProfileEditViewModel
     public FFmpegProfileVideoFormat VideoFormat { get; set; }
     public string VideoProfile { get; set; }
     public string VideoPreset { get; set; }
+    public bool AllowBFrames { get; set; }
     public FFmpegProfileBitDepth BitDepth { get; set; }
 
     public CreateFFmpegProfile ToCreate() =>
@@ -75,6 +77,7 @@ public class FFmpegProfileEditViewModel
             VideoFormat,
             VideoProfile,
             VideoPreset,
+            AllowBFrames,
             BitDepth,
             VideoBitrate,
             VideoBufferSize,
@@ -102,6 +105,7 @@ public class FFmpegProfileEditViewModel
             VideoFormat,
             VideoProfile,
             VideoPreset,
+            AllowBFrames,
             BitDepth,
             VideoBitrate,
             VideoBufferSize,